+++ app/upgrades/mainnet/v3/upgrades.go
@@ -1,0 +1,63 @@
+package v3
+
+import (
+	"github.com/cosmos/cosmos-sdk/client"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/genutil/types"
+	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
+
+	feemarketv010 "github.com/tharsis/ethermint/x/feemarket/migrations/v010"
+	feemarketv09types "github.com/tharsis/ethermint/x/feemarket/migrations/v09/types"
+	feemarkettypes "github.com/tharsis/ethermint/x/feemarket/types"
+)
+
+const UpgradeName = "v3"
+
+// CreateUpgradeHandler creates an SDK upgrade handler for v3
+func CreateUpgradeHandler(
+	mm *module.Manager,
+	configurator module.Configurator,
+) upgradetypes.UpgradeHandler {
+	return func(ctx sdk.Context, _ upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
+		// Refs:
+		// - https://docs.cosmos.network/master/building-modules/upgrade.html#registering-migrations
+		// - https://docs.cosmos.network/master/migrations/chain-upgrade-guide-044.html#chain-upgrade
+
+		// migrate fee market module, other modules are left as-is to
+		// avoid running InitGenesis.
+		vm[feemarkettypes.ModuleName] = 1
+
+		return mm.RunMigrations(ctx, configurator, vm)
+	}
+}
+
+// MigrateGenesis migrates exported state from v2 to v3 genesis state. It performs a no-op if the migration errors.
+func MigrateGenesis(appState types.AppMap, clientCtx client.Context) types.AppMap {
+	// Migrate x/feemarket.
+	if appState[feemarkettypes.ModuleName] == nil {
+		return appState
+	}
+
+	// unmarshal relative source genesis application state
+	var oldFeeMarketState feemarketv09types.GenesisState
+	if err := clientCtx.Codec.UnmarshalJSON(appState[feemarkettypes.ModuleName], &oldFeeMarketState); err != nil {
+		return appState
+	}
+
+	// delete deprecated x/feemarket genesis state
+	delete(appState, feemarkettypes.ModuleName)
+
+	// Migrate relative source genesis application state and marshal it into
+	// the respective key.
+	newFeeMarketState := feemarketv010.MigrateJSON(oldFeeMarketState)
+
+	feeMarketBz, err := clientCtx.Codec.MarshalJSON(&newFeeMarketState)
+	if err != nil {
+		return appState
+	}
+
+	appState[feemarkettypes.ModuleName] = feeMarketBz
+
+	return appState
+}